refactor: update images link and interfaces
