+++ src/components/error-boundary-component/error-boundary-component.module.css
@@ -7,7 +7,7 @@
   max-width: 1920px;
   width: 100vw;
   height: 100vh;
-  background-image: url("https://mrwallpaper.com/images/hd/darth-vader-4k-blue-lightsaber-x1o7b80w9sr1b9iv.jpg");
+  background-image: url("/background-error-boundary.jpg");
   background-repeat: no-repeat;
   background-position: center;
   background-size: cover;
